scraper_project/requirements.txt
mongoengine~=0.29.1
python-dotenv~=1.0.1
requests~=2.32.3
beautifulsoup4~=4.12.3